Windows以外でコンパイルエラーになりそうな部分を修正

+++ Dev/FontGenerator/FontGeneratorCore/Generator.cpp
@@ -22,7 +22,7 @@ namespace FontGenerator
 	{
 		SetNkfOption((char*)mode);
 
-		DWORD returnedBytes = 0;
+		unsigned long returnedBytes = 0;
 
 		NkfConvertSafe(outBuffer, outBufferSize, &returnedBytes, (char*) inStr, inStrlength);
 	}
